--- agnpy/__init__.py
@@ -1,6 +1,0 @@
-from .spectra import *
-from .emission_regions import *
-from .synchrotron import *
-from .compton import *
-from .targets import *
-from .absorption import *
