+++ agnpy/compton/external_compton.py
@@ -1,0 +1,1 @@
+# module containing the Externañ Compton radiative process
