+++ agnpy/emission_regions/__init__.py
@@ -1,0 +1,1 @@
+from .blob import *
